USB/COM/RFM22B: Added a standard com callback to test for link availability.  This replaces the PIOS_USB_CheckAvailable function with a generic PIOS_COM_Available function.  This is now used by the RFM22B com driver to test if a good link is up.  This was originally written by James, and Brian merged it into the latest RFM22B branch.

+++ flight/PiOS/STM32F4xx/pios_usb.c
@@ -155,7 +155,7 @@ int32_t PIOS_USB_ChangeConnectionState(bool connected)
  * \return 0: interface not available
  */
 uint32_t usb_found;
-bool PIOS_USB_CheckAvailable(uint8_t id)
+bool PIOS_USB_CheckAvailable(uint32_t id)
 {
 	struct pios_usb_dev * usb_dev = (struct pios_usb_dev *) pios_usb_id;
